Network-scale cumulative effects assessments

Assessment of cumulative effects and related metrics using the Beauchesne et al. 2021 method.

ncea(
  drivers,
  vc,
  sensitivity,
  metaweb,
  trophic_sensitivity,
  w_d = 0.5,
  w_i = 0.25,
  motif_summary = NULL
)

cea_binary(effect)

triads(metaweb)

cea_pathways(effect, vc)

ncea_pathways(vc_id, motifs)

ncea_pathways_(direct_pathways, motifs)

ncea_motifs(direct_effect, indirect_pathways)

ncea_effects(motif_summary, w_d = 0.5, w_i = 0.25)

get_species_contribution(motif_effects)

get_direct_indirect(motif_effects)

get_net(motif_effects)

get_cekm(motif_effects, vc)

Arguments

drivers

distribution and intensity of environmental drivers as stars object

vc

distribution of valued components as stars object

sensitivity

matrix of environmental drivers and valued component, with same name as those used in drivers and vc

metaweb

matrix of valued component by valued component describing the binary interations structuring the network of valued components

trophic_sensitivity

data.frame of trophic sensitivities, default from Beauchesne. Available as data package with data(trophic_sensitivity)

motif_summary

list of parameters obtained from ncea_motifs to allow for faster assessments, since that part of the assessment is the longest to run.

exportAs

string, the type of object that should be created, either a "data.frame" or a "stars" object

Functions

cea_binary(): transform effects assessment into binary 2D matrix to assess the presence of an effect to a valued component in a specific grid cell

triads(): assess all triads of interest from metaweb

cea_pathways(): pathways of direct effect

ncea_pathways(): pathways of indirect effect and trophic sensitivity

ncea_pathways_(): apply ncea_pathways to get pathways of indirect effect and trophic sensitivity for all cells

ncea_motifs(): get effects of drivers for species in all motifs in each cell

ncea_effects(): evaluate effects for all motifs using trophic sensitivity and effect weights

get_species_contribution(): get contribution of species to indirect effects

get_direct_indirect(): get direct and indirect effects of drivers

get_net(): get net effects of drivers

get_cekm(): get effects per km2

Examples

# Data
drivers <- rcea:::drivers 
vc <- rcea:::vc
sensitivity <- rcea:::sensitivity
metaweb <- rcea:::metaweb
trophic_sensitivity <- rcea::trophic_sensitivity
trophic_sensitivity$Sensitivity <- trophic_sensitivity$Sensitivity2
#> Warning: Unknown or uninitialised column: `Sensitivity2`.

if (FALSE) {
# Network-scale effects 
beauchesne <- ncea(drivers, vc, sensitivity, metaweb, trophic_sensitivity)
plot(beauchesne$net)
plot(beauchesne$direct)
plot(beauchesne$indirect)
}