Network-scale cumulative effects assessments

Assessment of cumulative effects and related metrics using the Beauchesne et al. 2021 method.

ncea_(
  drivers,
  vc,
  sensitivity,
  metaweb,
  trophic_sensitivity,
  w_d = 0.5,
  w_i = 0.25,
  vc_focus,
  direct_effect = NULL,
  motif_summary = NULL
)

Arguments

drivers

distribution and intensity of environmental drivers as stars object

vc

distribution of valued components as stars object

sensitivity

matrix of environmental drivers and valued component, with same name as those used in drivers and vc

metaweb

matrix of valued component by valued component describing the binary interations structuring the network of valued components

trophic_sensitivity

data.frame of trophic sensitivities, default from Beauchesne. Available as data package with data(trophic_sensitivity)

motif_summary

list of parameters obtained from ncea_motifs to allow for faster assessments, since that part of the assessment is the longest to run.

exportAs

string, the type of object that should be created, either a "data.frame" or a "stars" object

Functions

ncea_(): effects assessment using the Beauchesne approach, i.e. assessment of direct and indirect effects of drivers on valued components (species), with method applied on a single species at a time for ease of parallelisation

Examples

# Data
drivers <- rcea:::drivers 
vc <- rcea:::vc
sensitivity <- rcea:::sensitivity
metaweb <- rcea:::metaweb
data(trophic_sensitivity)
#> Warning: data set ‘trophic_sensitivity’ not found

if (FALSE) {
# Network-scale effects 
beauchesne <- ncea(drivers, vc, sensitivity, metaweb, trophic_sensitivity)
plot(beauchesne$net)
plot(beauchesne$direct)
plot(beauchesne$indirect)
}