Network-scale cumulative effects assessments for a single focal species

Assessment of cumulative effects and related metrics using the Beauchesne et al. 2021 method.

Usage

ncea_species(
  focus,
  drivers,
  vc,
  sensitivity,
  metaweb,
  trophic_sensitivity,
  w_d = 0.5,
  w_i = 0.25,
  output = "output/ncea",
  output_format = "geotiff"
)

Arguments

focus

named argument, string with name of valued component on which to perform the assessment.

drivers

distribution and intensity of environmental drivers as stars object

vc

distribution of valued components as stars object

sensitivity

matrix of environmental drivers and valued component, with same name as those used in drivers and vc

metaweb

matrix of valued component by valued component describing the binary interations structuring the network of valued components

trophic_sensitivity

data.frame of trophic sensitivities, default from Beauchesne. Available as data package with data(trophic_sensitivity)

w_d, w_i

weight for the direct (w_d) and indirect (w_i) modules when calculating network-scale cea scores; w_d + 2*w_i should be equal to 1.

output

relative path to export results of assessment.

output_format

output format, one of geotiff or COG.

Examples

# Data
drivers <- rcea:::drivers
vc <- rcea:::vc
sensitivity <- rcea:::sensitivity
metaweb <- rcea:::metaweb
trophic_sensitivity <- rcea::trophic_sensitivity

if (FALSE) {
# Network-scale effects for individual species
ncea_species(focus = "vc1", drivers, vc, sensitivity, metaweb, trophic_sensitivity)
ncea_species(focus = "vc1", drivers, vc, sensitivity, metaweb, trophic_sensitivity, output_format = "COG")
}