Simulate example 3-state data

Simulate example multi-state data for a 3-state model using Weibull distributions for each transition. Patients are right censored after 15 years of followup.

sim_mstate3_data(n = 2000, seed = NULL)

Arguments

n
The number of patients to simulate.

seed
An integer to pass to set.seed(). If NULL, then no seed is set.

Value

Returns a data.table with multiple rows for patient. That is, for a given patient there are rows for each possible transition from a given health state. The returned columns are as follows:

from

The health state ID transitioned from.

to

he health state ID transitioned to.

female

1 if female and 0 if male.

age

The age of the patient in years.

patient_id

The patient ID.

final

An indicator equal to 1 if a patient is in their final health state during the simulation and 0 otherwise.

time_start

The time at the start of the interval.

time_stop

The time and the end of the interval.

time

The duration of the interval: time_stop - time_start.

status

1 if time is an observed transition and 0 if there was right censoring.

transition_id

The transition ID representing a unique transition from from to to.

strategy_id

The treatment strategy ID.

strategy_name

The name of the treatment strategy. Either "New or "SOC.

Examples

sim_mstate3_data(n = 3, seed = 101)

#>    from to female      age patient_id final time_start time_stop status
#> 1:    1  2      0 61.17898          1     0   0.000000  3.939316      1
#> 2:    1  3      0 61.17898          1     0   0.000000  3.939316      0
#> 3:    2  3      0 61.17898          1     1   3.939316 12.589138      1
#> 4:    1  2      0 61.70923          2     0   0.000000  8.533489      1
#> 5:    1  3      0 61.70923          2     0   0.000000  8.533489      0
#> 6:    2  3      0 61.70923          2     1   8.533489 15.000000      0
#> 7:    1  2      1 66.45681          3     0   0.000000 15.000000      0
#> 8:    1  3      1 66.45681          3     0   0.000000 15.000000      0
#>    transition_id strategy_id strategy_name      time
#> 1:             1           2           New  3.939316
#> 2:             2           2           New  3.939316
#> 3:             3           2           New  8.649822
#> 4:             1           1           SOC  8.533489
#> 5:             2           1           SOC  8.533489
#> 6:             3           1           SOC  6.466511
#> 7:             1           1           SOC 15.000000
#> 8:             2           1           SOC 15.000000